Update `UnityLogListener` implementation to handle uncaught exceptions
`LogException` is called by Unity when an exception is uncaught, and has its own short call stack; to get the stack of the original thrown exception, `new StackTrace(exception)` is used and passed up through the log listener. In non-exception logs, the usual flow of `new StackFrame(...)` is used to resolve the caller context.