Add PhysicsUtil
Add a method to extrapolate the state of a rigid body by a delta without needing to step Unity's physics system.

Approximates the clean movement of a rigid body accounting for gravity, linear drag and angular drag. Does not account for collisions or any changes applied through the FixedUpdate of MonoBehaviours. However the latter can be accomplished by providing an onStep callback and calling the same logic that any FixedUpdate behaviours do.

Note: That values aren't re-read off of the RigidBody so all changes in onStep must make changes to the state parameters passed into the ExtrapolateState method.